sct: implement simple query service for epochs
