+++ crates/core/component/sct/src/lib.rs
@@ -6,11 +6,12 @@ pub mod component;
 
 pub mod event;
 pub mod genesis;
+pub mod params;
+pub mod state_key;
+
 mod nullification_info;
 mod nullifier;
-pub mod params;
 mod source;
-pub mod state_key;
 
 pub use nullification_info::NullificationInfo;
 pub use nullifier::{Nullifier, NullifierVar};
